checking if file alreayd exists in ressource downloader

+++ download_resources.py
@@ -11,15 +11,18 @@ def download_and_extract_dataset():
 
     file_path = os.path.join(save_dir, file_name)
 
-    print(f'downloading file from url : {url} ')
-    response = requests.get(url)
+    if not os.path.exists(resources_folder + file_name):
+        print(f'downloading file from url : {url} ')
+        response = requests.get(url)
 
-    if response.status_code == 200:
-        with open(file_path, 'wb') as file:
-            file.write(response.content)
-        print(f"File downloaded and saved to {file_path}")
+        if response.status_code == 200:
+            with open(file_path, 'wb') as file:
+                file.write(response.content)
+            print(f"File downloaded and saved to {file_path}")
+        else:
+            print(f"Failed to download the file. Status code: {response.status_code}")
     else:
-        print(f"Failed to download the file. Status code: {response.status_code}")
+        print(f'{resources_folder + file_name} already existes, no downloading required')
 
     bash_command = "tar -xvf " + resources_folder + file_name + " -C " + resources_folder
     print(f'executing : {bash_command}')
